Merge remote-tracking branch 'origin/caf/caf-wlan/master' into jb_mr1

* origin/caf/caf-wlan/master:
  wlan: Revision 3.2.3.273
  HDD : Increase private data length in hdd ioctl
  wlan: Fix build for FEATURE_WLAN_CH_AVOID
  wlan: dot11f.frms file
  wlan: Fix compilation errors.
  wlan : Revision 3.2.3.272
  CSA: MAC: Channel Switch Announce based on Beacon IE
  CSA: WDA: Channel Switch Announce based on Beacon IE
  wlan: make cfg.dat file read 64 bit compatible
  wlan : Revision 3.2.3.271
  prima : Handling race condition in driver unload in FTM
  wlan : Revision 3.2.3.270
  wlan: Add ini parametr to configure dwell time for first time
  wlan: Logging Infrastructure
  wlan: fix typecasting pointer as 32 bit in nv_parser.
  wlan: fix format specifier in logs.
  wlan: Remove dependency on obsolete mpp.h header file.

+++ CORE/SYS/common/src/wlan_qct_sys.c
@@ -204,8 +204,8 @@ typedef struct sPolFileHeader
 {
   tPolFileVersion FileVersion;
   tPolFileVersion HWCapabilities;
-  unsigned long   FileLength;
-  unsigned long   NumDirectoryEntries;
+  unsigned int   FileLength;
+  unsigned int   NumDirectoryEntries;
 
 } tPolFileHeader;
 
@@ -224,9 +224,9 @@ typedef enum ePolFileDirTypes
 
 typedef struct sPolFileDirEntry
 {
-  unsigned long DirEntryType;
-  unsigned long DirEntryFileOffset;
-  unsigned long DirEntryLength;
+  unsigned int DirEntryType;
+  unsigned int DirEntryFileOffset;
+  unsigned int DirEntryLength;
 
 } tPolFileDirEntry;
